+++ core-ng/src/main/java/core/framework/api/module/SessionConfig.java
@@ -44,7 +44,10 @@ public class SessionConfig {
         }
 
         logger.info("create redis session provider, host={}", host);
-        Redis redis = new RedisBuilder().host(host).get();
+        Redis redis = new RedisBuilder()
+            .host(host)
+            .poolSize(8, 32)      // reasonable value for AWS medium/large instances
+            .get();
         context.shutdownHook.add(redis::shutdown);
         context.httpServer.siteManager.sessionManager.sessionProvider(new RedisSessionStore(redis));
         return this;